OAID: remove use of waitForKeyElements()
waitForKeyElements uses a continuous 300ms interval. That's a
substantial system burden when the user might have hundreds or
thousands of tabs open. The browser will limit actions in
inactive tabs, but it's still substantial and unnecessary
under most circumstances. On SE, for what we're looking for
here, new post menus, those happen only in response to
AJAX fetches, so we hook $(document).ajaxComplete(),
so we're only looking when there *might* have been
a change that we're interested in.