OAID: For revisions, use a button instead of link.
This is consistent with SE's use of `<button>` vs `<a>` for these
types of buttons where the button isn't actually a link. There
is, in some views of the revisions pages, at least one other
button in the list that uses `<button>`, rather than `<a>`.

As it was, the `<a>` was causing the page to scroll to the top,
so the alternative is to prevent the default action. OTOH,
I found it annoying to have the button use link coloring and
change color on all of all "Detect OpenAI" links on the
revisions page once one of the links had been clicked and "visited".
While the visited color change would be fixed by preventing the
default, it wouldn't fix the link coloring being used for something
that's really not a link.